services/api/requirements.txt
Django==5.0.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-environ==0.11.2
psycopg2-binary==2.9.9
gunicorn==22.0.0
argon2-cffi==23.1.0
django-cors-headers==4.4.0
drf-spectacular==0.27.2
redis==5.0.7
dj-database-url==2.2.0

